identity: update context logger with 'device_id' or 'user_id' fields
Add support for updating context logger with 'user_id' or 'device_id' fields
based on the identity data. This assumes that the tokens have 'mender.device' or
'mender.user' claims. If none of these claims are found, the middleware will add
'sub' field to the logger, setting its value to the value of 'sub' claim.

Changelog: title

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>