global: avoid use of basic types as context.Context keys
The rationale behind this is that basic types are accessible everywhere, so it
is possible that someone may accidentally use the same value as the key and
overwrite existing Value in the context. A recommended solution for this is
using package private types. According to Go's equality rules:

  Interface values are comparable. Two interface values are equal if they have
  identical dynamic types and equal dynamic values or if both have value nil.

Since Key in context.Context is an interface{}, use of package private types
with either contstant non-nil value or value set at runtime will be sufficient.

Changelog: title

Signed-off-by: Maciej Borzecki <maciej.borzecki@rndity.com>